refactor: joinCode Context로 변경
