+++ frontend/src/App.tsx
@@ -1,10 +1,10 @@
 import { ThemeProvider } from '@emotion/react';
 import { Outlet } from 'react-router-dom';
-import { ModalProvider } from './components/@common/Modal/ModalContext';
-import { theme } from './styles/theme';
-import { PlayerRoleProvider } from './contexts/PlayerRole/PlayerRoleProvider';
 import { WebSocketProvider } from './apis/websocket/contexts/WebSocketProvider';
+import { ModalProvider } from './components/@common/Modal/ModalContext';
 import { JoinCodeProvider } from './contexts/JoinCode/JoinCodeProvider';
+import { PlayerRoleProvider } from './contexts/PlayerRole/PlayerRoleProvider';
+import { theme } from './styles/theme';
 
 const App = () => {
   return (
