+++ frontend/src/contexts/JoinCode/JoinCodeProvider.tsx
@@ -1,16 +1,11 @@
-import { JoinCodeContext } from './JoinCodeContext';
 import { PropsWithChildren, useState } from 'react';
+import { JoinCodeContext } from './JoinCodeContext';
 
 export const JoinCodeProvider = ({ children }: PropsWithChildren) => {
-  const [joinCode, setJoinCodeState] = useState<string | null>(null);
+  const [joinCode, setJoinCodeState] = useState<string>('');
 
-  const setJoinCode = (code: string) => {
-    setJoinCodeState(code);
-  };
-
-  const clearJoinCode = () => {
-    setJoinCodeState(null);
-  };
+  const setJoinCode = (code: string) => setJoinCodeState(code);
+  const clearJoinCode = () => setJoinCodeState('');
 
   return (
     <JoinCodeContext.Provider value={{ joinCode, setJoinCode, clearJoinCode }}>
